Use logging fixtures from the conftest
